Revert "More error handling and test cases"
