+++ api/idr.py
@@ -1,7 +1,7 @@
 from statistics import mean
 from math import sqrt
 
-from api.utils import get_score_std, get_sorted_responses, get_item_ids, get_student_list, update_input, get_error
+from api.utils import get_score_std, get_sorted_responses, get_item_ids, get_student_list, update_input
 from api.config import get_service_config, get_keyword_value
 
 
@@ -23,10 +23,9 @@ def calculate_idr(param):
              item ids as keys and idr as values
     """
     service_key = get_service_config(2)
-    catch_error = get_error(param)
-    if catch_error[0]:
-        return {service_key: catch_error[1]}
     inp = update_input(param)
+    if inp == get_keyword_value("no_students"):
+        return {service_key: get_keyword_value("no_students")}
     sorted_resp = get_sorted_responses(inp)
     num_students = len(sorted_resp)
     num_items = len (sorted_resp[0])
@@ -90,10 +89,9 @@ def calculate_idr_average(param):
     :return: a float: the average idr
     """
     service_key = get_service_config(11)
-    catch_error = get_error(param)
-    if catch_error[0]:
-        return {service_key: catch_error[1]}
     inp = update_input(param)
+    if inp == get_keyword_value("no_students"):
+        return {service_key: get_keyword_value("no_students")}
     idr_dict = list(calculate_idr(inp).values())[0]
     if idr_dict == get_keyword_value("bad_mean"):
         return {service_key: get_keyword_value("bad_mean")}
