+++ api/scores.py
@@ -1,4 +1,4 @@
-from api.utils import get_sorted_responses, update_input, get_student_ids, get_error
+from api.utils import get_sorted_responses, update_input, get_student_ids
 from api.config import get_service_config, get_keyword_value
 
 
@@ -14,10 +14,9 @@ def calculate_scores(param):
              student ids as keys and their score as values
     """
     service_key = get_service_config(4)
-    catch_error = get_error(param)
-    if catch_error[0]:
-        return {service_key: catch_error[1]}
     inp = update_input(param)
+    if inp == get_keyword_value("no_students"):
+        return {service_key: get_keyword_value("no_students")}
     sorted_resp = get_sorted_responses(inp)
     student_ids = get_student_ids(inp) 
     num_items = len (sorted_resp[0])
@@ -48,10 +47,9 @@ def calculate_average(param):
     :return: a float: the score average
     """
     service_key = get_service_config(5)
-    catch_error = get_error(param)
-    if catch_error[0]:
-        return {service_key: catch_error[1]}
     inp = update_input(param)
+    if inp == get_keyword_value("no_students"):
+        return {service_key: get_keyword_value("no_students")}
     score_list = list(list(calculate_scores(inp).values())[0].values())
     num_students = len(score_list)
     average = sum(score_list) / num_students
