+++ tests/test_functions.py
@@ -342,7 +342,7 @@ class TestFunctions:
     # testing the assumptions
     def test_assumptions(self):
 
-        expected = {"1234": {"assumed": ["7"]}, "1235": {"assumed": ["7"]}}
+        expected = {'1234': {'assumed': ['7']}, '1235': {'assumed': ['7']}}
         assumption = get_assumptions(self.data)["assumptions"]
 
         assert assumption == expected
@@ -370,3 +370,184 @@ class TestFunctions:
         topic_rights = calculate_topic_averages(self.data)["topic_avgs"]
 
         assert topic_rights == expected
+
+    # testing with no grad year
+    def test_no_grad_year(self):
+        data = {
+            "student_list": [
+                {
+                  "id": 1,
+                  "item_responses": [
+                        {"item_id": 1, "response": 1},
+                        {"item_id": 2, "response": 0},
+                    ]
+                },
+                { 
+                  "id": 2,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 1},
+                    ]
+                },
+                { 
+                  "id": 3,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 1},
+                    ]
+                }
+            ]
+        }
+
+        expected = "No graduation years were found, or all students are in the same graduation year"
+        analysis = analyze_grad_years(data)["grad_year_analysis"]
+
+        assert analysis == expected
+
+    # testing item exclude calcs
+    def test_item_excludes_difference(self):
+        data_1 = {
+            "student_list": [
+                {
+                  "id": 1,
+                  "item_responses": [
+                        {"item_id": 1, "response": 1},
+                        {"item_id": 2, "response": 0},
+                        {"item_id": 3, "response": 1},
+                        {"item_id": 4, "response": 0}
+                    ]
+                },
+                { 
+                  "id": 2,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 1},
+                        {"item_id": 3, "response": 0},
+                        {"item_id": 4, "response": 1}
+                    ]
+                },
+                { 
+                  "id": 3,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 0},
+                        {"item_id": 3, "response": 1},
+                        {"item_id": 4, "response": 1}
+                    ]
+                },
+                { 
+                  "id": 4,
+                  "item_responses": [
+                        {"item_id": 1, "response": 1},
+                        {"item_id": 2, "response": 1},
+                        {"item_id": 3, "response": 0},
+                        {"item_id": 4, "response": 0}
+                    ]
+                }
+            ],
+            "exclude_items": [4]
+        }
+        data_2 = {
+            "student_list": [
+                {
+                  "id": 1,
+                  "item_responses": [
+                        {"item_id": 1, "response": 1},
+                        {"item_id": 2, "response": 0},
+                        {"item_id": 3, "response": 1},
+                        {"item_id": 4, "response": 0}
+                    ]
+                },
+                { 
+                  "id": 2,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 1},
+                        {"item_id": 3, "response": 0},
+                        {"item_id": 4, "response": 1}
+                    ]
+                },
+                { 
+                  "id": 3,
+                  "item_responses": [
+                        {"item_id": 1, "response": 0},
+                        {"item_id": 2, "response": 0},
+                        {"item_id": 3, "response": 1},
+                        {"item_id": 4, "response": 1}
+                    ]
+                },
+                { 
+                  "id": 4,
+                  "item_responses": [
+                        {"item_id": 1, "response": 1},
+                        {"item_id": 2, "response": 1},
+                        {"item_id": 3, "response": 0},
+                        {"item_id": 4, "response": 0}
+                    ]
+                }
+            ],
+            "exclude_items": [1,3]
+        }
+
+        idr_1 = calculate_idr(data_1)
+        idr_2 = calculate_idr(data_2)
+
+        assert idr_1 != idr_2
+
+    # test analysis with no grad year, student id, or item id
+    def test_optional_inputs(self):
+        data = {
+            "student_list": [
+                {
+                  "item_responses": [
+                        {"response": 1},
+                        {"response": 0},
+                        {"response": 0}
+                    ]
+                },
+                { 
+                  "item_responses": [
+                        {"response": 0},
+                        {"response": 1},
+                        {"response": 1}
+                    ]
+                },
+                { 
+                  "item_responses": [
+                        {"response": 0},
+                        {"response": 1},
+                    ]
+                }
+            ]
+        }
+
+        expected = {"analysis": {
+                        "assumptions": {'3': {'assumed': ['3']}},
+                        "average": 44.4,
+                        "diff_avg": 0.556,
+                        "difficulty": {"1": 0.667,
+                                    "2": 0.333,
+                                    "3": 0.667},
+                        "exclude": ["1"],
+                        "grad_year_analysis": "No graduation years were found, or all students are in the same graduation year",
+                        "idr": {"1": -0.037,
+                                "2": 0.037,
+                                "3": 0.074},
+                        "idr_avg": 0.025,
+                        "kr20": -3.0,
+                        "num_correct": {"1": 1,
+                                        "2": 2,
+                                        "3": 1},
+                        "scores": {"1": 33.3,
+                                "2": 66.7,
+                                "3": 33.3},
+                        "topic_avgs": "No topics were found",
+                        'topic_rights': 'No topics were found',
+                        "weighted_avg": 40.0,
+                        "weighted_scores": {"1": 40.0,
+                                            "2": 60.0,
+                                            "3": 20.0}},
+                    }
+        analysis = analyze_test(data)
+
+        assert analysis == expected
